北京2021&2022空气质量历史数据处理
This is a small project to calculate the annual mean value of atmospheric monitoring data in Beijing. The data is sourced from 北京市环境保护检测中心网站(http://www.bjmemc.com.cn/), which is updated daily.
The result is a table showing the DAILY mean values of CO/NO2/O3/SO2 for the years 2021 and 2022.